[FIX] account_invoice_production_lot: handle return picking. These changes are intended to cover the following use case scenario.
Steps to reproduce:

* create a new product 'Product Test': stockable, Invoicing Policy as Delivered quantities, and tracking by lots
* update its quantity to 4 with new lot 'Lot1'
* update its quantity to 4 with new lot 'Lot2'
* sale 4 of 'Product Test', confirm the SO
* on the delivery, split the lots and set To Do quantities as 4 for Lot1
* validate the delivery
* return all quanties related to 'Lot 1'
* return of a return and use 'Lot 2' instead of 'Lot 1'
* validate the new delivery

Current behavior:

the field 'Production Lots' of the invoice generated is filled with both lots and, consequently,
the invoice report will also contain info related to the lot not delivered yet.

Expected behavior:

the invoice report only contains info related to the lot involved in the new delivery, which is 'Lot 2'.